Aetheris-main/aetheris-backend/requirements.txt
# Aetheris Backend — Python Dependencies
# Install: pip install -r requirements.txt

# ── CORE FRAMEWORK ────────────────────────────────────────────────────────
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.7.0

# ── DATABASE ──────────────────────────────────────────────────────────────
sqlalchemy==2.0.36
aiosqlite==0.20.0              # SQLite async driver (dev/hackathon)
asyncpg==0.30.0                # PostgreSQL async driver (production)
alembic==1.14.0                # DB migrations

# ── HTTP CLIENT ───────────────────────────────────────────────────────────
httpx==0.28.1                  # Async HTTP (OpenFDA, Whisper API calls)

# ── AI / LLM ──────────────────────────────────────────────────────────────
anthropic==0.40.0              # Claude API
openai==1.58.1                 # Whisper speech-to-text

# ── MACHINE LEARNING ──────────────────────────────────────────────────────
scikit-learn==1.5.2            # ML models (ASA risk, complication predictor)
numpy==1.26.4                  # Numerical computing
pandas==2.2.3                  # Data manipulation

# ── CACHING ───────────────────────────────────────────────────────────────
redis==5.2.1                   # Redis client (optional for production)

# ── SECURITY ──────────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0   # JWT tokens
passlib[bcrypt]==1.7.4             # Password hashing

# ── UTILITIES ─────────────────────────────────────────────────────────────
python-multipart==0.0.20       # File/form uploads
python-dotenv==1.0.1           # .env file loading

# ── TESTING ───────────────────────────────────────────────────────────────
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1                  # Test client
groq==0.13.0
